% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_closest_dist.R
\name{calc_closest_dist}
\alias{calc_closest_dist}
\title{Calculate Closest Distance and Time}
\usage{
calc_closest_dist(track, geography, geoid = "GEOID")
}
\arguments{
\item{geography}{\code{sf} object for the area of interest}

\item{geoid}{Unique identifier for geography}

\item{hurdat2}{Output from \code{get_hurdat2()} or interpolated track from
\code{calc_interpolated_track()}}
}
\value{
Closest distance and time to track for each geography
}
\description{
Calculates at what time point the storm is closest and records the distance
from track at that time.
}
\examples{
\dontrun{
harvey_closest_dist <- calc_closest_dist(
  harvey_hurdat2,
  TXLA_tracts
)
}
}
