% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_closest_dist.R
\name{calc_interpolated_track}
\alias{calc_interpolated_track}
\title{Calculate Interpolated Track}
\usage{
calc_interpolated_track(track)
}
\arguments{
\item{track}{Output from \code{get_hurdat2()}}
}
\value{
Interpolated track at 15 minute intervals
}
\description{
Utilizes \href{https://github.com/geanders/stormwindmodel}{\code{stormwindmodel}}
to impute the \code{hurdat2} track to 15 minute intervals.
}
